requirements.txt
Flask>=2.2
Flask-SQLAlchemy
Flask-Migrate
Flask-Marshmallow
marshmallow-sqlalchemy
flasgger
gunicorn